Move django-static.ts into django.ts.
